# improve performance, improve typings