--- a/.github/workflows/node.js.yml
+++ b/.github/workflows/node.js.yml
@@ -13,4 +13,7 @@ on:
 jobs:
   test:
     uses: fastify/workflows/.github/workflows/plugins-ci.yml@v3
+    with:
+      lint: true
+