--- a/benchmark/secret.js
+++ b/benchmark/secret.js
@@ -49,4 +49,4 @@ suite.on('complete', function onComplete () {
   benchmarks.log()
 })
 
-suite.run({ async: false })
+suite.run({ async: false, delay: 0 })