--- a/index.js
+++ b/index.js
@@ -1,3 +1,5 @@
+'use strict'
+
 /*!
  * csrf
  * Copyright(c) 2014 Jonathan Ong
@@ -6,34 +8,7 @@
  * MIT Licensed
  */
 
-'use strict'
-
-/**
- * Module dependencies.
- * @private
- */
-
-const rndm = require('rndm')
-const uid = require('uid-safe')
-const compare = require('tsscmp')
 const crypto = require('crypto')
-
-/**
- * Module variables.
- * @private
- */
-
-const EQUAL_GLOBAL_REGEXP = /=/g
-const PLUS_GLOBAL_REGEXP = /\+/g
-const SLASH_GLOBAL_REGEXP = /\//g
-const MINUS_GLOBAL_REGEXP = /-/g
-
-/**
- * Module exports.
- * @public
- */
-
-module.exports = Tokens
 
 /**
  * Token generation/verification class.
@@ -86,6 +61,7 @@ function Tokens (options) {
   }
 
   this.saltLength = saltLength
+  this.saltGenerator = saltGenerator(saltLength)
   this.secretLength = secretLength
   this.validity = validity
   this.userInfo = userInfo
@@ -110,7 +86,7 @@ Tokens.prototype.create = function create (secret, userInfo) {
     }
   }
 
-  return this._tokenize(secret, rndm(this.saltLength), date, userInfo)
+  return this._tokenize(secret, this.saltGenerator(), date, userInfo)
 }
 
 /**
@@ -120,41 +96,149 @@ Tokens.prototype.create = function create (secret, userInfo) {
  * @public
  */
 
-Tokens.prototype.secret = function secret (callback) {
-  return uid(this.secretLength, callback)
-}
+Tokens.prototype.secret = Buffer.isEncoding('base64url')
+  ? function secret (callback) {
+    // validate callback is a function, if provided
+    if (callback !== undefined && typeof callback !== 'function') {
+      throw new TypeError('argument callback must be a function')
+    }
+
+    // require the callback without promises
+    if (!callback && !global.Promise) {
+      throw new TypeError('argument callback is required')
+    }
+
+    if (callback) {
+      // classic callback style
+      crypto.randomBytes(this.secretLength, (err, buf) => {
+        err
+          ? callback(err)
+          : callback(null, buf.toString('base64url'))
+      })
+      return
+    }
+
+    return new Promise((resolve, reject) => {
+      crypto.randomBytes(this.secretLength, (err, buf) => {
+        err
+          ? reject(err)
+          : resolve(buf.toString('base64url'))
+      })
+    })
+  }
+  : function secret (callback) {
+    // validate callback is a function, if provided
+    if (callback !== undefined && typeof callback !== 'function') {
+      throw new TypeError('argument callback must be a function')
+    }
+
+    // require the callback without promises
+    if (!callback && !global.Promise) {
+      throw new TypeError('argument callback is required')
+    }
+
+    if (callback) {
+      // classic callback style
+      return crypto.randomBytes(this.secretLength, function (err, buf) {
+        err
+          ? callback(err)
+          : callback(null, buf
+            .toString('base64')
+            .replace(PLUS_GLOBAL_REGEXP, '-')
+            .replace(SLASH_GLOBAL_REGEXP, '_')
+            .replace(EQUAL_GLOBAL_REGEXP, ''))
+      })
+    }
+
+    return new Promise((resolve, reject) => {
+      crypto.randomBytes(this.secretLength, (err, buf) => {
+        err
+          ? reject(err)
+          : resolve(buf
+            .toString('base64')
+            .replace(PLUS_GLOBAL_REGEXP, '-')
+            .replace(SLASH_GLOBAL_REGEXP, '_')
+            .replace(EQUAL_GLOBAL_REGEXP, ''))
+      })
+    })
+  }
 
 /**
  * Create a new secret key synchronously.
  * @public
  */
 
-Tokens.prototype.secretSync = function secretSync () {
-  return uid.sync(this.secretLength)
-}
+Tokens.prototype.secretSync = Buffer.isEncoding('base64url')
+  ? function secretSync () {
+    return crypto.randomBytes(this.secretLength)
+      .toString('base64url')
+  }
+  : function secretSync () {
+    return crypto.randomBytes(this.secretLength)
+      .toString('base64')
+      .replace(PLUS_GLOBAL_REGEXP, '-')
+      .replace(SLASH_GLOBAL_REGEXP, '_')
+      .replace(EQUAL_GLOBAL_REGEXP, '')
+  }
 
 /**
  * Tokenize a secret, salt, date and userInfo.
  * @private
  */
 
-Tokens.prototype._tokenize = function tokenize (secret, salt, date, userInfo) {
-  let toHash = ''
+Tokens.prototype._tokenize = Buffer.isEncoding('base64url')
+  ? function tokenize (secret, salt, date, userInfo) {
+    let toHash = ''
 
-  if (date !== null) {
-    toHash += date.toString(36) + '-'
+    if (date !== null) {
+      toHash += date.toString(36) + '-'
+    }
+
+    if (typeof userInfo === 'string') {
+      // we hash the userInfo to ensure it's encoded properly and to have a fixed length
+      userInfo = crypto
+        .createHash('sha1')
+        .update(userInfo)
+        .digest('base64url')
+        .replace(MINUS_GLOBAL_REGEXP, '_')
+      toHash += userInfo + '-'
+    }
+
+    toHash += salt
+
+    return toHash + '-' + crypto
+      .createHash('sha1')
+      .update(toHash + '-' + secret, 'ascii')
+      .digest('base64url')
   }
+  : function tokenize (secret, salt, date, userInfo) {
+    let toHash = ''
 
-  if (typeof userInfo === 'string') {
-    // we hash the userInfo to ensure it's encoded properly and to have a fixed length
-    userInfo = hash(userInfo).replace(MINUS_GLOBAL_REGEXP, '_')
-    toHash += userInfo + '-'
+    if (date !== null) {
+      toHash += date.toString(36) + '-'
+    }
+
+    if (typeof userInfo === 'string') {
+      // we hash the userInfo to ensure it's encoded properly and to have a fixed length
+      userInfo = crypto
+        .createHash('sha1')
+        .update(userInfo)
+        .digest('base64')
+        .replace(MINUS_SLASH_GLOBAL_REGEXP, '_')
+        .replace(EQUAL_GLOBAL_REGEXP, '')
+      toHash += userInfo + '-'
+    }
+
+    toHash += salt
+
+    return toHash + '-' + crypto
+      .createHash('sha1')
+      .update(toHash + '-' + secret, 'ascii')
+      .digest('base64')
+      .replace(PLUS_GLOBAL_REGEXP, '-')
+      .replace(SLASH_GLOBAL_REGEXP, '_')
+      .replace(EQUAL_GLOBAL_REGEXP, '')
   }
-
-  toHash += salt
-
-  return toHash + '-' + hash(toHash + '-' + secret)
-}
 
 /**
  * Verify if a given token is valid for a given secret.
@@ -175,21 +259,21 @@ Tokens.prototype.verify = function verify (secret, token, userInfo) {
   }
 
   let index = token.indexOf('-')
-  const toCompare = token
-  let date = null
-
   if (index === -1) {
     return false
   }
 
+  const actual = Buffer.from(token)
+  let date = null
+
   if (this.validity > 0) {
-    date = parseInt(token.substr(0, index), 36)
+    date = parseInt(token.slice(0, index), 36)
 
     if (Date.now() - date > this.validity) {
       return false
     }
 
-    token = token.substr(index + 1)
+    token = token.slice(index + 1)
     index = token.indexOf('-')
 
     if (index === -1) {
@@ -206,7 +290,7 @@ Tokens.prototype.verify = function verify (secret, token, userInfo) {
 
     // we skip the userInfo part, this will be
     // verified with the hashing
-    token = token.substr(index + 1)
+    token = token.slice(index + 1)
     index = token.indexOf('-')
 
     if (index === -1) {
@@ -214,24 +298,38 @@ Tokens.prototype.verify = function verify (secret, token, userInfo) {
     }
   }
 
-  const salt = token.substr(0, index)
-  const expected = this._tokenize(secret, salt, date, userInfo)
+  const salt = token.slice(0, index)
+  const expected = Buffer.from(this._tokenize(secret, salt, date, userInfo))
 
-  return compare(toCompare, expected)
+  // to avoid the exposure if the provided value has the right length, we call
+  // timingSafeEqual with the actual value. The length check itself is timing
+  // safe
+  return crypto.timingSafeEqual(
+    actual.length === expected.length
+      ? expected
+      : actual,
+    actual
+  ) && actual.length === expected.length
 }
 
-/**
- * Hash a string with SHA1, returning url-safe base64
- * @param {string} str
- * @private
- */
+const EQUAL_GLOBAL_REGEXP = /=/g
+const PLUS_GLOBAL_REGEXP = /\+/g
+const SLASH_GLOBAL_REGEXP = /\//g
+const MINUS_GLOBAL_REGEXP = /-/g
+const MINUS_SLASH_GLOBAL_REGEXP = /[-/]/g
 
-function hash (str) {
-  return crypto
-    .createHash('sha1')
-    .update(str, 'ascii')
-    .digest('base64')
-    .replace(PLUS_GLOBAL_REGEXP, '-')
-    .replace(SLASH_GLOBAL_REGEXP, '_')
-    .replace(EQUAL_GLOBAL_REGEXP, '')
+function saltGenerator (saltLength) {
+  const fnBody = []
+
+  fnBody.push('const base62 = \'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789\'.split(\'\');')
+  fnBody.push('return function () {')
+  const salt = []
+  for (let i = 0; i < saltLength; ++i) salt.push('base62[(62 * Math.random()) | 0]')
+  fnBody.push('return ' + salt.join('+'))
+  fnBody.push('}')
+  return new Function(fnBody.join(''))() // eslint-disable-line no-new-func
 }
+
+module.exports = Tokens
+module.exports.default = Tokens
+module.exports.Tokens = Tokens